refactor: 입력 값을 검증 및 변환하는 책임을 InputView 에 할당하기.
